move #7 : directory 수정
