+++ src/main/java/smwu/heartcall/domain/emergency/entity/Emergency.java
@@ -1,4 +1,4 @@
-package smwu.heartcall.domain.emergencyCall.entity;
+package smwu.heartcall.domain.emergency.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
