+++ src/main/java/smwu/heartcall/domain/emergency/service/EmergencyService.java
@@ -1,15 +1,15 @@
-package smwu.heartcall.domain.emergencyCall.service;
+package smwu.heartcall.domain.emergency.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import smwu.heartcall.domain.emergencyCall.dto.CallEmergencyRequestDto;
-import smwu.heartcall.domain.emergencyCall.dto.EmergencyAudioUrlResponseDto;
-import smwu.heartcall.domain.emergencyCall.dto.EmergencyDetailResponseDto;
-import smwu.heartcall.domain.emergencyCall.dto.GetEmergencyResponseDto;
-import smwu.heartcall.domain.emergencyCall.entity.Emergency;
-import smwu.heartcall.domain.emergencyCall.repository.EmergencyRepository;
+import smwu.heartcall.domain.emergency.dto.CallEmergencyRequestDto;
+import smwu.heartcall.domain.emergency.dto.EmergencyAudioUrlResponseDto;
+import smwu.heartcall.domain.emergency.dto.EmergencyDetailResponseDto;
+import smwu.heartcall.domain.emergency.dto.GetEmergencyResponseDto;
+import smwu.heartcall.domain.emergency.entity.Emergency;
+import smwu.heartcall.domain.emergency.repository.EmergencyRepository;
 import smwu.heartcall.domain.user.entity.User;
 import smwu.heartcall.domain.user.repository.RelationRepository;
 import smwu.heartcall.domain.user.repository.UserRepository;
